Do not special case seal event response handling.
In an attempt to debug the XON interception issue the handling of a
response by the quixote manager to seal a security domain was handled by
its own case statement.  Since the only action required is to process
the 'OK' command from the Sancho implementation this special casing was
removed.