Implement ERR macro if one has not been defined.
The MCU Sancho implementation cannot use the standard ERR macro since
the printf statement it uses would corrupt the TTYduct packet
processing.  The MCU implementation provides its own implementation of
the ERR macro so this update checks for the definition of an ERR macro
and only provides one if it is absent.