Implement single buffer transmission in MCU TTYduct.
In an attempt to debug the XON interception problem tha was previously
discussed and addressed the write of a TTYduct buffer was split into two
pieces.

The length of the packet was first sent followed by the actual packet
conents themselves.  This update simplifies that and simply sends the
entire contents of the Buffer object after it has been assembled.