Implement tracing to an output file.
Previously setting the -d command-line argument caused tracing output to
be sent to stderr.  These changes require the specification of a
command-line argument for the -d switch and writes trace output to that
file.  This simplifies the ability to review the dialogue between
Quixote and the Sancho sub-system in use.