Enable cellular alert support.
This update implements support for transmitting disciplining alerts from
a cellular enabled Sancho micro-controller implementation.  The base
cellular support was a part of the initial commit, this update finalizes
the implementation and allows it to be enabled through the
quixote-console interface.

The following command was added by this commit to the quixote-console
command syntax:

enable cellular

This causes the Sancho instance to turn on the cellular radio system
that in turn causes a data connection to be established for the
endpoint.

When the directive to enable cellular operations is received, and
after the cell service thread is started, the command interpreter loops
for 30 seconds waiting for Cell Services Task (CST) to return
CST_MODEM_DATA_READY_STATE.  This is the automata state that indicates
that packet based cellular connection is active and the modem has
received a network IP address.

If this state is returned in 30 seconds a value of 'OK' is returned to
Quixote, otherwise a FAILED string is returned.

This update also adds a guard enumeration named as follows:

sancho_cmds_max

The code that verifies the correctness of the command code now uses that
value, a situation that prevents having to modify this check when
additional commands may be added.