Add utility to test a threaded application in a security domain.
This utility adds a simple utility named test-thread for testing the
behavior of a threaded application in a security event domain.

The primary process spawns a thread that pauses for five seconds and
then opens the /etc/group file and prints its contents in hexadecimal
format.

The primary process waits for 10 seconds and then opens the /etc/passwd
file and prints the contents of that file in hexadecimal format.