Implement automatic reset for SanchoMCU implementation.
The current implemenation of an MCU based Sanco evaluator is a simple
single session implementation.  As a result the micro-controller needs
to be reset after each cartridge session in order for the Quixote
implementation to synchronize its communications with the device.

The following command syntax was added to the micro-controller command
interpreter to instruct the card to issue a system reset and return it
to its normal power up status:

reset

The quixote utility was modified to send this command when a cartridge
exit is detected.  This greatly simplifies iteractive execution and
testing of a Quixote/SanchoMCU session.