Fix security state loading.
The -m switch to a quixote instance specifies that the file specified as
an argument to the -m switch should be read and processed in order to
setup the security state of a Sancho instance.  This updated fixes a
number of issues with the initial implementation of this command.

The statically file scoped instance of the TTYduct variable was removed.
The code has moved toward a model where the TTYduct variable is simply
passed as an argument to the callers that need it.

The enumeration type the command to set a specific security state was
changed from 'contour_event' to 'sancho_state' in a move toward
prefixing all of the enumeration states with the sanch_ preamble.

The nomenclature change was propagaged into the Sancho-MCU instance as
well.

The function to read the security domain state definition was changed
from:

initialize_model

To:

initialize_state

The previous implementation had left the terminating newline on the
initialize directive.  The file read routine was modified to convert
that newline to a null character.