Implement user space quixote modeling.
This commit adds an initial implementation of a quixote instance
that implements state modeling in a userspace process.

This is a first approximation and is notable for the fact that
the state and measurement values of a security state domain do
not match the value from the execution trajectory from which the
states were dumped.