Update set-behavior utility.
This commit begins the process of modifying the set-behavior utility to
be consistent with the new security domain nomenclature.  References to
the 'ai' prefix are all changed to the corresping 'te' prefix.

The -e command-line switch was added in order to access the new
functionality added to the sys_config_events_ns system call that allows
all of the Turing Event to be set to enforcing mode where they return
-EPERM if accessed by a process with bad actor status.