Remove support for specifying external modeling.
By definition the quixote utility will be implementing external security
domain modeling so the notion of allowing this to be specified by a
command-line switch is incorrect.

Remove the command-line switch and have the setup_namespace code
explicitly write the value required to seal the security domain into the
pseudo-file that manages this functionality.

This commit also modifies the quixote-us.c source to fix an issue with
the comment indicating that the following code is for defining the name
of the update file since the previous comment was out of place.