Drop CAP_TRUST bounding after domain setup.
The CAP_TRUST capability is not needed in the process that is running
the security domain after the configuration of the domain.  This update
drops the capability from the bounding set so that subordinate processes
lose their ability to monitor the trust state of the domain.

This results in the current process still having the CAP_TRUST
capability when the runc utility is invoked.  Additional investigation
needs to be done to see if a preferable strategy would be to remove the
CAP_TRUST capability in all three froms from the current process
capability set.