Updates to system call and setup_namespace.
The defines for the system call numbers and arguments were somewhat
rationalized.  These need to migrate into an independent header file so
that there is not a need to keep multiple locations synchronized.

The setup_namespace() function was modified to use a system call to
specify that the kernel should use external modeling rather then writing
to the /sysfs pseudo-file.

The call to the clone() function to create a new security domain was
modified to return an error status rather then print an error string so
that it is consistent with most of the other lower level functions.

The add_event() function uses the new system call definitions for better
clarity.