Fix system call issues in Sancho-MCU implementation.
Testing of the MCU based Sancho implementation indicated a regression
when a call was made to set the actor status of a process.  This was
traced to an incorrect system call number.  This update fixes the system
call number.

Further testing indicated that an incorrect flag value was being used in
the call to release a process in 'good actor' status.  The correct
define used to set this flag was also adjusted in this commit.