Allow quixote-us to return an error from process_event().
The process_event() function in the Quixote-US implementation was
unconditionally setting the return value of the function to true,
thus disregarding a potential error return values from the
specific event handling routines.

This would cause the Quixote-US implementation to spin in a loop
if an error was encountered during event processing.