Improve termination handling.
The current quixote implementations handled termination poorly.
The most problematic issue was leaving the runc supervisor and
the resulting process stack stranded in the background if the
quixote launch process terminated.

This was introduced by introducing the kill_cartridge() function
that is called whenever the quite launch process terminates.
This function runs the followingn command:

runc kill CARTRIDGE SIGKILL

Where CARTRIDGE is replaced with the name of the software
cartridge that is being run.  This causes runc to send the
SIGKILL process to the init process of the domain which results
in the termination of the domain and the subsequent cleanup of
the runc infrastructure.

This commit also introduces some standardization and improvements
in naming convention.

This commit also sets the debug file, if activated, into line
mode buffering.