Standardize on using 'goto' to exit pool loop.
The code clause that handles an error return from the primary
quixote pool loop was using a 'break' instruction to leave the
pool loop.  All of the other loop exits were using a goto call to
branch to the program exit handling.  Replace the 'break' with
the appropriate goto in order to have a standardized
implementation.