Properly handle exit of a poisoned model.
The security domain modeling object was generating a segmentation
fault when it was called after an object poisoning event.  This
was caused by the fact that the exit clause was attempting to use
the ContourPoint object that would have been allocated in the
event of normal event processing.

Conditionalize the exit code to update the discipline and
addition status of the point only if update processing has been
executed.