Add release mode to bad-actor utility.
Debugging the Quixote utilities can lead to situations where a
security domain gets 'stuck' secondary to the fact that the
modeling process has failed.  In this case it is useful to have a
method for setting the hung process to 'good actor' status so the
Quixote utility can properly clean up the execution environment.

This commit adds the '-r PID' command-line argument that sets the
status of the process identified by the PID arguement to be a
good actor.

The system call that implements that is gated by the need to have
the CAP_TRUST capability so this utility can only be run from a
process that carries that capability.