Clean up properly after a modeling engine failure.
If an error occurs in the modeling engine processes in a security
domain that are being supervised cannot make forward progress.
This creates issues with respect to the ability to have runc
properly cleap up the security domain properly.

If this situation is encountered the modeler is placed into a
configuration that approves of all security state events and runc
is signaled to terminate the domain.

Further investigation needs to be conducted to determine if
additional safeguards can be placed into effect to reduce the
security state events to the minimal subset required to implement
the shutdown of the domain.