Implement support for setuid execution.
In order to support non-privileged execution this commit modifies
the codepath that spawns a security domain specific process to
set the uid of the process to the real process uid of the
effective uid does not match the real uid.   This allows the
quixote-us and quixote-mcu utilities to be installed setuid root
which is necessary in order to support creation of the namespace.

The setup_management() function was modified to allow the
management sockets directory to be writable by members of the
group that own the directory.  This allows the Quixote instances
started by a non-privileged user to be managed by that user as
long as they are a member of the group that owns the sockets
management directory.