Change nomenclature for showing security event states.
The command to display the set of unique points in the security
domain was previously the following:

show contours

This was secondary to older nomenclature that had proven to be
somewhat unclear.

This commit changes this command to the following:

show points

The preference for this would be to use the command 'show states'
but that does not currently work with the method that is used to
parse commands.  This syntax may be reconsidered at a later time.