Fix show mode in quixote-mcu.
The change to the use of an enumeration type to specify the mode
that the utilities are to run in omitted the need to specify the
static variable for the conditional that checked for 'show mode'.
This was causing the inability of the quixote-mcu utility to show
the available software cartridges.

This commit fixes that conditional so that show mode now works
properly.