Fix quixote-mcu segmentation fault.
The quixote-mcu utility was generating a segmentation fault when
run in runc mode and no cartridge was specified.  This was
secondary to an attempt to dereference the Duct object in order
to send a reset signal the co-processor.

Fix this by explicitly testing for a valid Duct pointer before
sending the reset command.