Implement management of multiple Quixote instances.
The previous implementation of Quixote was simplistic and
implemented only a single UNIX domain socket location.  This
proved to be problematic in cases where multiple security domains
were run on a single platform.

This update addresses this issue and allows multiple security
domains and types to be hosted.

The following locations are used for the management sockets:

/var/run/Quixote/cartridges

/var/run/Quixote/processes

The former directory contains the management sockets for security
domains that are being in a full alternate namespace environment.
The latter directory contains the sockets for security domains
that are running in process mode.

The -S command-line switch was added to quixote-console to allow
the currently active security domains to be displayed.

The -p and -c command-line switches were added to allow the
specification of a particular security domain.  The argument to
the -p command-line argument is the process id (PID) of the
security domain instance.  The argument to the -c command-line
switch is the name of the software cartridge being run.