Implement security domain mapping support for in-kernel quixote.
This update implements the ability for quixote to load a security
state map into the kernel for in-kernel discipling of a domain.

Since the updates to the kernel security domain need to be
conducted from the namespace that owns the domain, a different
strategy is needed for quixote as compared to the instances that
use external Sancho advisors in order to avoid the 'Heisenberg
Effect'.

The 'Heisenberg Effect' in trusted systems is the principle that
a definition for a trusted security domain cannot be measured by
the domain that is being configured.  As a result the quixote
utility has to open the state map in the parent event namespace
and pass the file descriptor of the file to the namespace that
will configure itself.

In the processing of the state map the 'aggregate' event
directive is disregarded since the security domain has already
had an aggregate measurement injected into it.  Support was added
for adding states points to the security domain as well as the
sealing of the domain.