Modify quixote-console output based on input/output type.
The quixote-console utility would always generate a command-line
prompt even if it was receiving its command from a pipe.  Since
the command prompt was being generated to stderr this was causing
a series of prompts to be output if commands were being piped
into the utility over stdin.

The utility was also modified to determine if output was going to
a terminal or to a device/pipe.  If output is going to a pipe or
device, a count of the number of output lines, in the case of
points, trajectory or forensics, are not output.

In the case of output to a terminal, the utility now defaults to
line buffered output so flush commands are not needed in
interactive mode in order to have immediate output when in
interactive mode.