Implement support for output of security state map.
This update provides an initial implementation of the ability to
generate a security state map that can be used to define the
behavior of subsequent invocations of a security event namespace.

This map can be injected into a new security domain by using the
'-m STATE.MAP' command-line option to the quixote utilities.
Where STATE.MAP is replaced with the name of the file containing
the security state directives.

The following command option was added to the quixote-console
utility to initiate the output of a security state map:

show map

A map file is of the following format:

aggregate AGGREGATE_VALUE
state STATE_VALUE
...
seal

Where AGGREGATE_VALUE and STATE_VALUE are replaced with the
ASCII hexadecimal encoding of the platform hardware aggregate
value and the ASCII hexadecimal encoding of a state point in
the event namespace.

The seal command is provided in order to simplify the consistency
of the generation of a map and is always output.

The update provides an implementation for the quixote-us and
quixote-mcu implementations.  The in-kernel quixote
implementation requires some additional support infrastructure in
the Quixote kernel driver.  That implementation will be added
after the kernel support for obtaining the aggregate value
is added to the Quixote kernel driver.