Implement one-shot execution of quixote-console commands.
In order to facilitate better non-interactive use of the
quixote-console tool the commit enables the ability to execute
most management commands through the specification of a
command-line switch.

The following command-line arguments were added as subsitutions
for the quixote commands:

-E -> show events

-F -> show forensics

-M -> show map

-P -> show points

-S -> show state

-T -> show trajectory

An equivalent for the 'show measurement' command was not
implemented, primarily because the -M option was used for the
'show map' command and a good alterntative command-line option
did not spring to mind for the 'show measurement' command.

Arguably the state value of a security domain is the more useful
description of a security domain then the measurement value.