Migrate SGX based modeling code to Sancho heirarchy.
This update largely represents a code movement that migrates the
implementation of the security state modeling code from the
following directory:

SRDE/ISOidentity

To the following location:

Sancho/SGX

Where it will be subsequently re-architected into the SanchoSGX
implementation.

The only code changes implemented were what was necessary to
obtain a clean compile of the source tree.