Initial implementation of SanchoSGX trusted modeling agent.
This commit provides an initial implementation of the quixote-sgx
utility that uses an SGX based Sancho implementation as its
Trusted Modeling Agent.

The current implementation lacks the following capabilities:

- Display of state value.

- Display of security state points.

- Display of TE events.

Additional infrastructure work needs to be done on the SanchoSGX
enclave in order to support output of those security domain
parameters.

The SGX enclave needed substantial updates to its OCALL
infrastructure in order to support enhancements that have been
made to the SRDEnaaaim libraries.

Notable was the movement of the OCALL enumerations and structure
out of the SanchoSGX object and into the enclave interface
infrastructure.  Additional work is needed on this infrastructure
in order to faciliate less troublesome expansion of the OCALL
infrastructure.

The current quixote-sgx implementation only allows interrogation
of the security domain state through the UNIX domain
quixote-console socket.  Implementation of the network based
interrogation is left to subsequent updates.