Convert enclave signing to srde-sign tool.
The signing of the SGX TMA implementation had previously been
done with the standard signing tool.  This commit converts the
signing process to use the srde-sign tool that is currently an
interface to the binary signing tool.

The XML based configuration files were converted to signing
meta-data file with a .meta file suffix.