Implement security domain model de-initialization.
The security domain state model was not being destroyed which was
resulting in a message being generated at enclave destruction
that there were a large number of objects in the trusted space
that were still valid when the enclave exited.

The argument structure for ECALL0 was modified to support a
boolean value that indicates whether or not the security state
model should be initialized or 'whacked'.

The SanchoSGX->whack() method was modified to issue ECALL0 with
this argument set to false in order to have the ECALL
de-initialize the model.

The init_model() function in the enclave was modified to access a
boolean argument which selects whether the INIT or WHACK macro
were to be invoked on the model object.