Implement SanchoSGX state retrieval.
The initial implementation of the SGX based Sancho TMA was not
capble of retrieving the state information for the security
domain.  This commit adds the necessary support for doing that.

The enclave based get_measurement() function was overloaded so
that a type value could be specified.  Since the measurement and
state of a domain are similar entities that only vary with
respect to their source this allowed the retrieval of the state
information without having to introduce an additional ECALL.

The following method was addessed to the Sancho SGX TMA to
support retrieval of the state value:

SanchoSGX->get_state

This function and the SanchoSGX->get_measurement method were
modified to set the the type of value to be returned in the ECALL
argument structure.

Code was uncommented in the 'show_state' enumeration clause in
the process_command() function in quixote-sgx to enable
processing of the state request.

A dependency against the SanchoSGX object was added to the rule
to build the quixote-sgx binary so the binary would be
automatically rebuilt if the object was modified.