Implement method for trajectory size retrieval.
The ISOidentity->size() method was being used to retrieve the
number of points in the security domain model.  This was
problematic in the fact of situations where a security state
model was being loaded into the object.

The ->size method was returning the number of defined points in
the security state model.  Routines which were retrieving the
security event trajectory were using the ->size method to
interrogate the number of points that would be requested.

When the security state was being externally defined through its
state points this led to a situation where retrieval of the
domain trajectory would fail since the ->size method was
returning a value larger then the number of entries in the
trajectory list.

The size of the TE events and forensics list had already been
converted to independent methods.  In order to avoid any
confusion the ->trajectory_size method was added by this commit
in order to explicitly retrieve the number of elements in the
trajectory.

This sets the stage for deprecation of the ->size method
completely that will occur in a subsequent update.