Implement retrieval of security domain state points.
The SanchoSGX enclave implementation lacked the ability to return
the state points in a security domain.  This was largely due to
the fact that the legacy implementation was using a network based
SGX console agent for interrogating the model.

This update adds an additional ECALL to the enclave to allow
state points to be retrieved.  Consistent with the forensics and
trajectory retrieval this method is designed to be called
interatively for the number of points in the list.

The SanchoSGX object had the following method added to it:

->trajectory_size

In order to return the size of the execution trajectory list
explicitly.  Previously the ->size method was being used that
would cause an erroneous value to be returned if the security
domain model was being configured directly with state points.

Proper implementation of the security state points now allows a
state map to be generated from a SanchoSGX implementation.

Additional naming clarification needs to be conducted in the
security domain modeling object, currently ISOidentity, that will
be renamed in a subsequent update.