Fix loading of enclave from memory.
The SRDEenclave->load_memory method had suffered from code
neglect and was no longer capable of loading and initializing an
enclave from a memory buffer.  This was secondary to the fact
that this method was not properly dealing versioned metadata.

This commit updates the ->load_memory method to use the same code
infrastructure as the ->load method to support multiple versions
of metadata.