Make enclave include file dependent on enclave.
The implementation of the unified SanchoSGX implementation had
not made the header file that defines the array containing the
enclave dependent on the signed SanchoSGX-enclave.  As a result a
change in the enclave was not triggering a rebuild of the header
file.

This commit adds a Makefile dependency that will now force the
rebuild of the header file if the enclave changes.