Implement return of TE descriptions from SanchoSGX.
The SGX enclave implementing security domain modeling did not
have the ability to return the list of Trusted Event descriptions
generated by security state events that were not in the model
being implemented.

The TE_EVENTS define was added to the SanchoSGX-interface.h file
to provide a type description for ecalls that rewind, return the
size and elements of the TE list.

The get_event() function was modified to provide a code path that
will return a String value containing the TE description when the
TE_EVENTS type is specified.

The get_size() function was modified to return the number of TE
descriptions when called with the TE_EVENTS type.

The rewind() function was modified to reset the TE event cursor
when called with a TE_EVENTS type.

The following three methods were added to the SanchoSGX object in
order to support the retrieval of TE descriptions:

->get_te_event

->te_size

->rewind_te

The send_events() function in both the quixote-sgx and
quixote-sgx-u implementations were modified to actually implement
the return of the TE descriptions.  The new object methods were
used to implement the retrieval and return of the TE
descriptions.