Move identity infrastructure code to Attic.
The original sources committed were for the distributed identity
management system that had been designed.  No sub-directory
structure had been used so those files have historically remained
in the root of the source tree.

Since there is the change that this infrastructure may be of
potential use this commit moves them to the Attic directory for
the time being so they can be easily referenced.