Convert data sync/source object to new nomenclature.
The object that manages the characteristics of an inode or a data
sync or source is converted by this commit to the name of Cell.
This is in deference to the classic Turing model that uses the
term 'cell' to refer to locations on the infinite tape that
will be processed by the computing head.