Drop unused method in Cell object.
A following unused method has been languishing in the Cell object
for a long period of time:

->set_identity_elements

It was a remnant of when the object was not parsing a security
exchange event description.  The method has been commented out
for a long period of time and this commit drops it.