Update HurdLib.
This commit updates the HurdLib support library to a newer
development version.

This update enables access to the following new functionality:

The Gaggle object that allows arrays of objects to be managed.
This infrastructure will be used to replace the open coded
implementation of this in the TSEM object.

The String object now has support for adding or concantenating a
string based on the generation of a string from snprintf.