Drop ->size method from TSEM object.
The sizes of the various model components are now individually
accounted for by the components themselves and the previously
generic 'size' attribute for the object is no longer relevent.

This commit drops the internal state variable as well as its
query method.  In addition the use of the object was corrected in
multiple locations to use the component specific method call.