Convert TSEM modeling to use Gaggles's.
The commit converts the TSEM object from using an open coded
implementation of an array of objects to the Gaggle object.

At this commit level a modified HurdLib library is needed that
has a definition of the GPTR macro which should be as follows:

If compiling at this commit level the above macro needs to be
added to the TSEM.c file.

This commit adds the Gaggle implementation to the SRDEfusion
library that does have the GPTR macro implementation.