Initial support for security event modeling.
This commit adds the following object:

EventModel

Which is designed to implement algorithmic modification of the
parameters that define a security interaction event.  This allows
the final security state to be functionally directed to a
specific value.

The first secuirty model paradigm implement by this object is to
implement pseudonyms.  A pseudonym is a file that has a constant
digest value associated with it.  Pseudonyms are designed to
address files that are designed to be modified and are thus not
amenable to classic deterministic trust models.

Modeling of an event is implemented by invoking the
EventModel->evaluate method and passing it the SecurityEvent
object that is to be interpreted.  This object is designed to be
called before the ->measure method is called on the object.

The Cell->get_pseudonym method is implemented in order to compute
the pseudonym value for the Cell object.  The pseudonym value is
the value of the following function:

Pseudonym = SHA256(PATH_LENGTH || PATH_NAME)

This value is used to determine whether or not the digest value should
be set to zero if a subsequent Cell yields the same pseudonym value.

The Cell->set_digest method is implemented in order to set the digest
value of the Cell.

The SecurityEvent->evaluate_pseudonym method is implemented in order
to take a supplied pseudonym value and determine whether or not the
event should have a pseudonym applied to it.

The test-cell unit test was updated to print out the pseudonym value
for the test Cell object.

The test-event unit test was modified to evaluate the test event
against the known pseudonym value for the event in order to verify
that the pseudonym model is applied to the event.