Initial implementation of security model loading.
This commit provides the framework for the ability to load a
security model into a TSEM object.  The TSEM->load method was
added to the object to provide a dedicated method for loading the
model.

The first call to the ->load method sets a flag in the state data
of the object.  While this flag is set calls to the ->update
method will return an error.

The following directives are implemented in this first
instantiation of model loading:

aggregate	Defines the aggregate value for the model.

state		Defines a state point in the model.

pseudonym	Defines a pseudonym in the model.

seal		Prevents any further changes to the model.

end		Ends input of the security model.

Initial support for model loading is implemented by this commit
in the quixote-us utility.  Currently enforcement of the model is
not working since the variable that indicates the model is sealed
was being set by the event processing code.  This will be
addressed by returning the sealed status of the model in an
argument to the TSEM->update method.

The TSEM object now expects to call the SecurityEvent->measure
method rather then expecting to have the method be called by the
code that is feeding security events into the TSEM object.  This
is arguably a more correct and less error prone strategy.

The Quixote Makefile was not implementing dependencies on the
SecurityModel object code for the quixote utilities.  This update
implements this dependency so there is a clean rebuild of the
utilities when the SecurityModel code is updated.