Remove restriction on public key retrieval.
The RSAkey->get_public_key method was restricting the retrieval
of a public key from an object that either had a generated key or
had been loaded with a public key.  This restriced the retrieval
of a public key from an object that had been loaded with a
private key.

This was an artifical restriction since a generic private key can
always have a public key extracted from it so this commit removes
that restriction allowing a public key to be extracted from a
private key object.