Initial implementation of signed memory models.
This commit provides a basic implementation of the ability to
create a signed security model.  A signed model is indicated by
the following directive to start the model file:

key Base64_KEY

Where Base64_KEY is the base64 encoded public key that should be
used to verify the signature.

The signature is provided by the following directive which will
be the last directive in the file:

signature Base64_SIGNATURE

Where Base64_SIGNATURE is the base64 encoded signature of the
model file that is to be verified by the previously provided
public key.

This commit also introduces the following utility:

sign-model

Which is used to a signed security model map.

The command is invoked with the following syntax:

sign-model -k PRIVATE_KEY -m MODEL_FILE

Where:

	PRIVATE_KEY:	The name of the file containing the
			private key that will be used to sign
			the model.

	MODEL_FILE:	The name of the file containing a
			standard model definition that is to be
			signed.

The test-TSEM utility was updated to provide a method for testing
the loading of a security model.  The command is invoked in the
following form to set loading functionality:

test-TSEM -L -m model_file

The -L command-line argument selects for testing loading of a
security map and the -m command-line option is used to specify
the model file that is to be loaded.