Re-enable event disciplining.
This commit updates the quixote-us and quixote-sgx utilities so
that they properly discipline the process that was participating
in a security interaction event.

Proper disciplining was lost secondary to the updates to
implement the TSEM->load method.  This was secondary to the fact
that in the case of the loading of a security map via the ->load
method the Sealed static variable was not being set.

This was addressed by adding a boolean pointer to the
TSEM->update method.  The method sets the variable pointed to by
the pointer to indicate whether or not the model is sealed.  This
provides feedback on every update event with respect to the
sealed status of the model.

The quixote-sgx implementation still does not have support for
the ->load method since the SanchoSGX object needs to implement
an ECALL for the method.

The quixote-mcu implementation is currently not functional given
that we have no code support for Base64 encoding and RSA key
functionality.  This will be addressed in a subsequent commit
after full SGX functionality is implemented.