Remove TSEM->update_map method and enable SGX model loading.
This commit drops the TSEM method that had previously been used
to add security state points to a model.  All model
initialization is now implemented through the TSEM->load method.

The Sancho 'state' command is dropped since it is now longer used
to add security state points.

Security model loading was implemented in the quixote-sgx and
quixote-sgx-u utilities.  The ECALL for the SanchoSGX-enclave
that formerly implemented the ->update_map method was modified
and used to implement the ->load method.