Implement signed security maps for SanchoMCU.
This commit is a rather far ranging update whose predominent
intent is to implement the ability to load signed security maps
into a SanchoMCU instance.

Previously a 'one-off' implementation of the SHA256 hash
algorithm had been implementing using publically available source
code.  This commit implements the basic framework for
integratting cryptographic primitives from the mbdedtls
library.  The source code, as of this release, is reference built
against the 2.16.11 release which is the current 'long term
support' release.

Base64 and RSAkey objects were implemented with the mbedtls
primitives.  Sufficient support was implemented in the Base64
and RSAkey objects to support the interpretation of signed maps.

The 'load' command was implement in the sancho-cmds.h file to
allow commands to be submitted to the SanchoMCU instance that are
prefised with the 'load' verb.  The quixote-mcu utility was
modified to transmit the contents of a security map to the
SanchoMCU instance prefised with the 'load' directive.

Support was added to the SanchoMCU command interpreter to feed
any directives prefixed with the 'load' keyword to the TSEM->load
method.

The mbedtls_config.h file was added to the SanchoMCU implement.
This file specifies the directives that are needed to request the
compilation of the primitives that are needed in order to provide
the necessary cryptographic support for Base64 encoding and RSA
signature verification.

Infrastructure was added to provide a test harness for the
SanchoMCU implementation.  This was implemented in order to
provide a method for testing the cryptographic primitives that
were implemented in this commit.

The existing sha256.c and sha256.h files were renamed to the
following filenames:

sha256-conte.c

sha256-conte.h

In order to disambiguate the object files they were generating
from the objects files provided by the mbedtls implementation.