Implement NAAAIM library.
The cryptographic manipulation objects that were implemented in
the Sancho/STmcu/stm32l4xx/Sancho_Cell/Sancho directory are
objects which are implement in the lib directory of the NAAAIM
distribution.

This commit creates the following directory that will hold the
STmcu implementations of NAAAIM library files:

Sancho/STmcu/NAAAIM

In addition to creating the basic library infrastructure this
commit migrates the Base64 and RSAkey objects to being built as
the NAAAIM library objects.