Remove Sancho_Cell version of HurdLib header file.
The cellular enabled implementation of SanchoMCU was using an
independent implementation of the HurdLib master header file.  An
artifact from the initial implementation of the code.

This commit removes that header file and makes the build reliant
on the STmcu implementation of the HurdLib library.

Since the STmcu HurdLib implementation is a tiered implementation
it is important to note that the new Makefile has include
references for both the STmcu implementation of the library along
with the master library implementation.  The ordering of these
includes is important in that the STmcu implementation directory
must be searched first before the master implementation is
searched.

This insures that the STmcu specific implementation is used.
A subseqent update to the HurdLib library should make the use of
an implementation specific master header file unnecessary.